requirements.txt
appdirs==1.4.4
attrs==23.2.0
black==21.4b2
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
coverage==5.5
flake8==3.9.1
flake8-docstrings==1.6.0
Flask==2.1.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
MarkupSafe==2.1.5
mccabe==0.6.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
Pillow==9.5.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
pyproject-toml==0.0.10
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.9.0.post0
python-docx==0.8.10
pytz==2024.1
referencing==0.34.0
regex==2024.4.16
requests==2.25.1
requests-mock==1.9.2
rpds-py==0.18.0
setuptools-scm==6.0.1
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==2.2.2
